Push consumption metrics to server
For some reason when including the timestamp in the metric we push 5 copies get generated. Each copy is 1 minute further forward. So if we submitted a metric with the timestamp 16:00, we would get duplicate copies generated for 16:00, 16:01, 16:02, 16:03 and 16:04. If we use the start of the consumption hour bucket as the metric timestamp this means the duplicate copies are still within the correct hour bucket.