Don't include consumption data
There is an API quota for the consumption data that the script has been hitting. I could probably just limit the amount of time we request consumption data to once an hour, but the data we get from the API is no better than what's in the app, so not worth the hassle of working around.