Expose metrics via an endpoint using prometheus_client
Currently the script is both pushing metrics, and exposing metrics to be scraped. The pushed metrics are going to a public grafana cloud, and the scraped metrics will be used by a prometheus instance self hosted on the raspberry pi. Eventually I will set up remote_write to send metrics from the local prometheus to the public grafana.